NetPlay: Refactor some functions into a common header
